[Api][Zone] Enable unit tests
