--- disabledTests/Api/Admin/ZoneMembersTest.php
@@ -1,46 +1,0 @@
-<?php
-
-/*
- * This file is part of the Sylius package.
- *
- * (c) Sylius Sp. z o.o.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Sylius\Tests\Api\Admin;
-
-use Sylius\Component\Addressing\Model\ZoneMemberInterface;
-use Sylius\Tests\Api\JsonApiTestCase;
-use Sylius\Tests\Api\Utils\AdminUserLoginTrait;
-use Symfony\Component\HttpFoundation\Response;
-
-final class ZoneMembersTest extends JsonApiTestCase
-{
-    use AdminUserLoginTrait;
-
-    /** @test */
-    public function it_gets_zone_member(): void
-    {
-        $fixtures = $this->loadFixturesFromFiles(['authentication/api_administrator.yaml', 'zones.yaml']);
-        $header = array_merge($this->logInAdminUser('api@example.com'), self::CONTENT_TYPE_HEADER);
-
-        /** @var ZoneMemberInterface $zoneMember */
-        $zoneMember = $fixtures['member_NL'];
-
-        $this->client->request(
-            method: 'GET',
-            uri: sprintf('/api/v2/admin/zone-members/%s', $zoneMember->getId()),
-            server: $header,
-        );
-
-        $this->assertResponse(
-            $this->client->getResponse(),
-            'admin/zone_member/get_zone_member_response',
-            Response::HTTP_OK,
-        );
-    }
-}
