+++ tests/Api/Responses/admin/zone/create_zone.json
@@ -8,8 +8,17 @@
     "type": "province",
     "scope": "all",
     "members": [
-        "\/api\/v2\/admin\/zone-members\/@integer@",
-        "\/api\/v2\/admin\/zone-members\/@integer@",
-        "\/api\/v2\/admin\/zone-members\/@integer@"
+        {
+            "@type": "ZoneMember",
+            "code": "AL"
+        },
+        {
+            "@type": "ZoneMember",
+            "code": "CA"
+        },
+        {
+            "@type": "ZoneMember",
+            "code": "NY"
+        }
     ]
 }
