--- tests/Api/Responses/admin/zone/get_zone_members_response.json
@@ -1,22 +1,0 @@
-{
-    "@context": "\/api\/v2\/contexts\/ZoneMember",
-    "@id": "\/api\/v2\/admin\/zones\/EU\/members",
-    "@type": "hydra:Collection",
-    "hydra:member": [
-        {
-            "@id": "\/api\/v2\/admin\/zone-members\/@integer@",
-            "@type": "ZoneMember",
-            "id": "@integer@",
-            "code": "NL",
-            "belongsTo": "\/api\/v2\/admin\/zones\/EU"
-        },
-        {
-            "@id": "\/api\/v2\/admin\/zone-members\/@integer@",
-            "@type": "ZoneMember",
-            "id": "@integer@",
-            "code": "BE",
-            "belongsTo": "\/api\/v2\/admin\/zones\/EU"
-        }
-    ],
-    "hydra:totalItems": 2
-}
