+++ tests/Api/Responses/admin/zone/update_zone.json
@@ -8,7 +8,17 @@
     "type": "country",
     "scope": "all",
     "members": [
-        "\/api\/v2\/admin\/zone-members\/@integer@",
-        "\/api\/v2\/admin\/zone-members\/@integer@"
+        {
+            "@type": "ZoneMember",
+            "code": "NL"
+        },
+        {
+            "@type": "ZoneMember",
+            "code": "DE"
+        },
+        {
+            "@type": "ZoneMember",
+            "code": "PL"
+        }
     ]
 }
