--- tests/Api/Responses/admin/zone_member/get_zone_member_response.json
@@ -1,8 +1,0 @@
-{
-    "@context": "\/api\/v2\/contexts\/ZoneMember",
-    "@id": "\/api\/v2\/admin\/zone-members\/@integer@",
-    "@type": "ZoneMember",
-    "id": "@integer@",
-    "code": "NL",
-    "belongsTo": "\/api\/v2\/admin\/zones\/EU"
-}
